refactor: put some methods in a module for game_stats
