+++ lib/game_stats.rb
@@ -1,7 +1,8 @@
 require_relative 'stats'
+require_relative 'game_statables'
 
 class GameStats < Stats 
-  
+include GameStatables
   def initialize(files)
     super
   end
@@ -20,36 +21,12 @@ class GameStats < Stats
    all_home_wins.length.fdiv(@games.length).round(2)
   end
 
-  def all_home_wins
-    home_wins = []
-    @games.each do |game|
-      home_wins << game if game.home_goals.to_i > game.away_goals.to_i
-    end
-    home_wins
-  end
-
   def percentage_visitor_wins
     all_away_wins.length.fdiv(@games.length).round(2)
   end
 
-  def all_away_wins
-    away_wins = []
-    @games.each do |game|
-      away_wins << game if game.away_goals.to_i > game.home_goals.to_i
-    end
-    away_wins
-  end
-
   def percentage_ties
     all_ties.length.fdiv(@games.length).round(2)
-  end
-
-  def all_ties
-    tie_game = []
-    @games.each do |game|
-      tie_game << game if game.away_goals.to_i == game.home_goals.to_i
-    end
-    tie_game
   end
 
   def count_of_games_by_season
@@ -73,11 +50,4 @@ class GameStats < Stats
     end
   end
 
-  def goals_by_season
-    season_goals = Hash.new(0)
-    @games.each do |game|
-      season_goals[game.season_year] += game.total_score
-    end
-    season_goals
-  end
 end
